Add Vodafone-Ziggo
Adding some context on the brand names, which went like this iirc:

The Ziggo brand consists of a merger of two networks.

  * UPC (became part of Liberty-Global/AS6830)
  * Ziggo (former AS9143)

Ziggo/AS9143 merged with former UPC and became part of LG. Later
this was bought by Vodafone, creating Vodafone-Ziggo.

Former AS9143 has migrated to AS31615, the former UPC part has not.