Update operators.csv
Sorry to keep doing this but I messed up on the first two. Spectrum was listed twice, once as safe and once as unsafe. I ran the test myself on Charter's network and it came back unsafe. Additionally Spectrum is the brand name but the actual company is Charter Communications and that is the name the shows up when running the test and on DNS/APN searches (like whoismyisp.org) so I feel that name it he most appropriate for the list.